requirements.txt
# ── AI / Agents ───────────────────────────────────────
crewai==0.28.8
langchain==0.1.20
langchain-groq==0.1.3
langchain-community==0.0.38

# ── Vector store (resume RAG) ──────────────────────────
faiss-cpu==1.8.0

# ── Document parsing ───────────────────────────────────
pypdf==4.2.0
python-docx==1.1.2

# ── Web scraping ───────────────────────────────────────
requests==2.31.0
beautifulsoup4==4.12.3

# ── Backend ────────────────────────────────────────────
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1

# ── Database ───────────────────────────────────────────
sqlalchemy==2.0.30

# ── Frontend ───────────────────────────────────────────
streamlit==1.35.0

# ── Utils ──────────────────────────────────────────────
python-dotenv==1.0.1
tiktoken==0.7.0
